backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
anthropic==0.7.11
openpyxl==3.11.0
pandas==2.1.3
pdfplumber==0.10.2
pymupdf==1.23.5
reportlab==4.0.8
weasyprint==60.1
rapidfuzz==3.5.2
lxml==4.9.3
requests==2.31.0
